Prompt weighting; improve CFGZero*, rope, scheduled CFG; general cleanup + type hinting

+++ hv_generate_video.py
@@ -154,7 +154,7 @@ def save_images_grid(
 
 
 def save_videos_grid_prores(
-    videos: torch.Tensor, output_video: str, rescale: bool = False, fps: int = 24, n_rows: int = 1
+    videos: torch.Tensor, output_video: str, rescale: bool = False, fps: int = 24, n_rows: int = 1, keep_frames: bool = False
 ):
     videos = rearrange(videos, "b c t h w -> t b c h w")
     outputs = []
@@ -168,7 +168,7 @@ def save_videos_grid_prores(
         outputs.append(video)
     output_dir = os.path.dirname(output_video)
     os.makedirs(output_dir, exist_ok=True)
-    frame_dir = os.path.join(os.path.dirname(output_dir), "frames")
+    frame_dir = os.path.join(os.path.dirname(output_dir), "out_frames")
     os.makedirs(frame_dir, exist_ok=True)
     for idx, img in enumerate(outputs):
         image_path = os.path.join(frame_dir, f"{idx:04d}.png")
@@ -194,8 +194,8 @@ def save_videos_grid_prores(
         ],
         check=True
     )
-
-    shutil.rmtree(frame_dir)
+    if not keep_frames:
+        shutil.rmtree(frame_dir)
 
 # region Encoding prompt
 
@@ -544,6 +544,7 @@ def parse_args():
         help="Scale clip and llm influence"
     )
     parser.add_argument("--prores", action="store_true", help="Save video as Apple ProRes(perceptually lossless) instead of MP4")
+    parser.add_argument("--keep_frames", action="store_true", help="Keep intermediate frame directory(PNGs) when saving with prores")
     args = parser.parse_args()
     if args.cfg_schedule:
         args.cfg_schedule = parse_scheduled_cfg(args.cfg_schedule, args.infer_steps)
@@ -1059,7 +1060,7 @@ def main():
             sample = sample.unsqueeze(0)
             video_path = f"{save_path}/{time_flag}_{i}_{seeds[i]}{original_name}.mp4"
             if args.prores:
-                save_videos_grid_prores(sample, video_path.replace(".mp4", ".mkv"), fps=args.fps)
+                save_videos_grid_prores(sample, video_path.replace(".mp4", ".mkv"), fps=args.fps, keep_frames=args.keep_frames)
             else:
                 save_videos_grid(sample, video_path, fps=args.fps)
             logger.info(f"Sample save to: {video_path}")
